refactor: 클래스로 파일 분리
