+++ src/main/kotlin/blackjack/domain/result/DealerResult.kt
@@ -1,0 +1,5 @@
+package blackjack.domain.result
+
+import blackjack.domain.user.Dealer
+
+data class DealerResult(val dealer: Dealer, val winCount: Int, val drawCount: Int, val loseCount: Int)
