+++ src/main/kotlin/blackjack/domain/result/Result.kt
@@ -1,0 +1,5 @@
+package blackjack.domain.result
+
+enum class Result {
+    WIN, DRAW, LOSE
+}
