+++ src/main/kotlin/blackjack/domain/result/UserResult.kt
@@ -1,0 +1,5 @@
+package blackjack.domain.result
+
+import blackjack.domain.user.User
+
+data class UserResult(val user: User, val result: Result)
