+++ src/main/kotlin/blackjack/domain/result/UserResults.kt
@@ -1,0 +1,3 @@
+package blackjack.domain.result
+
+data class UserResults(private val userResults: List<UserResult>) : Iterable<UserResult> by userResults
