+++ src/main/kotlin/blackjack/domain/user/UserNames.kt
@@ -1,3 +1,3 @@
-package blackjack.domain
+package blackjack.domain.user
 
 data class UserNames(val nameSet: Set<String>) : Iterable<String> by nameSet
