+++ src/main/kotlin/blackjack/io/InputView.kt
@@ -1,6 +1,6 @@
 package blackjack.io
 
-import blackjack.domain.UserNames
+import blackjack.domain.user.UserNames
 
 object InputView {
     private const val USER_NAME_INPUT_MESSAGE = "게임에 참여할 사람의 이름을 입력하세요.(쉼표 기준으로 분리)"
