+++ src/test/kotlin/blackjack/domain/BlackjackResultTest.kt
@@ -1,5 +1,14 @@
 package blackjack.domain
 
+import blackjack.domain.card.Card
+import blackjack.domain.card.CardNumber
+import blackjack.domain.card.Cards
+import blackjack.domain.card.Suit
+import blackjack.domain.result.BlackjackResults
+import blackjack.domain.result.DealerResult
+import blackjack.domain.user.Dealer
+import blackjack.domain.user.User
+import blackjack.domain.user.Users
 import blackjack.util.TEST_USER_DRAW_INTERFACE
 import io.kotest.core.spec.style.BehaviorSpec
 import io.kotest.data.forAll
