+++ src/test/kotlin/blackjack/util/FixedCardsSelector.kt
@@ -1,8 +1,8 @@
 package blackjack.util
 
-import blackjack.domain.Card
-import blackjack.domain.Cards
-import blackjack.domain.Deck
+import blackjack.domain.card.Card
+import blackjack.domain.card.Cards
+import blackjack.domain.card.Deck
 
 class FixedCardsSelector(vararg cardList: Card) : CardSelector {
     private val deck = Deck(Cards(cardList.toList()))
