+++ src/test/kotlin/blackjack/util/TestUserDrawInterface.kt
@@ -1,5 +1,5 @@
 package blackjack.util
 
-import blackjack.domain.UserDrawInterface
+import blackjack.domain.user.UserDrawInterface
 
 val TEST_USER_DRAW_INTERFACE = UserDrawInterface { true }
